# inoxoft-task12/src/resources/controllers/user.controller.js
const { User } = require('../models');

const { AOUTH_FIELD_UR } = require('../../common/modelsFields.enum');
const emailActionsEnum = require('../../common/emailActions.enum');
const actionTypesEnum = require('../../common/actionTypes.enum');
const HttpStatusCode = require('../../common/statusCodes');
const { EM_FRONT_URL } = require('../../common/config');

const {
    authService,
    jwtService,
    userService,
    emailService
} = require('../services');

const asyncWrapper = require('../../middleware/asyncWrapper');

module.exports = {

    getAllUsers: asyncWrapper(async (req, res) => {
        const users = await userService.getAll(req.query);

        if (!users.data || users.data.length < 1) {
            return res.status(HttpStatusCode.NOT_FOUND).send({ message: 'No users found' });
        }

        return res.status(HttpStatusCode.OK).json({
            data: users.data.map(User.toResponse),
            page: users.page,
            limit: users.limit,
            count: users.count,
            pageCount: users.pageCount
        });
    }),

    getUserById: asyncWrapper(async (req, res) => {
        const { user_id } = req.params;

        const user = await userService.getUserById(user_id);

        if (!user) {
            return res.status(HttpStatusCode.NOT_FOUND).send({ message: 'User not found' });
        }

        return res.json(User.toResponse(user));
    }),

    createUser: asyncWrapper(async (req, res) => {
        const user = await userService.createUser(req.body);

        if (!user) {
            return res.status(HttpStatusCode.CONFLICT).send({ message: 'Can\'t create new User, try again' });
        }

        const actionToken = jwtService.generateActionToken(actionTypesEnum.FORGOT_PASS);

        await jwtService.createActionTokenInBd(actionToken, user._id);

        await emailService.sendMail(
            user.email,
            emailActionsEnum.CREATE_ACCOUNT,
            {
                login: user.email,
                userName: user.name,
                frontendResetPass: `${EM_FRONT_URL}/auth/password/forgot/set?token=${actionToken}`
            }
        );

        return res.status(HttpStatusCode.CREATED).send(`${user.role} created, email for reset password was sent`);

        // return res.status(HttpStatusCode.CREATED).json(User.toResponse(user));
    }),

    updateUser: asyncWrapper(async (req, res) => {
        const { user_id } = req.params;

        const user = await userService.updateUserById(user_id, req.body);

        if (!user) {
            return res.status(HttpStatusCode.NOT_FOUND).send({ message: 'User not found' });
        }

        // delete all tokens from BD with user._id
        await authService.userLogout(AOUTH_FIELD_UR, user_id);

        return res.status(HttpStatusCode.OK).json(User.toResponse(user));
    }),

    deleteUser: asyncWrapper(async (req, res) => {
        const { user_id } = req.params;

        const deletedUser = await userService.deleteUserById(user_id);

        if (!deletedUser) {
            return res.status(HttpStatusCode.NOT_FOUND).send({ message: 'User not found' });
        }

        // delete all tokens from BD with user._id
        await authService.userLogout(AOUTH_FIELD_UR, user_id);

        return res.status(HttpStatusCode.OK).send({ message: 'User has deleted' });
    }),
};